Improve failure reporting of update checks
The update checks can fail for various reasons, so it appears to be
appropriate to report why they failed.  We therefore throw exceptions,
instead of returning NULL or FALSE from hi_fsFileGetContents() and
hi_versionInfo(), and report the exception messages as failure reason.

Since update check failures are hopefully rare, we do not
internationalize the error messages, but rather hard-code them.  This
could still be improved, if necessary.

We neither report detailed errors for invalid local version.nfo, nor
for the quick update check.  The latter for obvious reasons, the former
might be improved later.

We throw general `RuntimeException`s for simplicity; this requires PHP
≥ 5.1.0, which doesn't appear to be an issue.